cohere-image-embeddings/requirements.txt
cohere
pillow
numpy 